fix some files
